cluster and cluster_info module added, check mode added to all modules

+++ plugins/modules/ssh_key.py
@@ -39,7 +39,9 @@ options:
     description: Keys marked as default are always included on server creation and reset unless toggled off in creation/reset request.
     type: bool
   ssh_key:
-    description: SSH Key actual key value.
+    description:
+      - SSH Key actual key value.
+      - Once a SSH Key is created, it cannot be modified through a playbook
     type: str
   name:
     description: Friendly SSH Key name to represent an SSH key.
@@ -113,12 +115,13 @@ ssh_key:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
-from ansible_collections.phoenixnap.bmc.plugins.module_utils.pnap import set_token_headers, HAS_REQUESTS, requests_wrapper, SSH_API
+from ansible_collections.phoenixnap.bmc.plugins.module_utils.pnap import set_token_headers, HAS_REQUESTS, requests_wrapper, check_immutable_arguments, SSH_API
 
 import os
 import json
 
 ALLOWED_STATES = ["present", "absent"]
+IMMUTABLE_ARGUMENTS = {'ssh_key': 'key'}
 
 
 def get_existing_keys(module):
@@ -135,28 +138,35 @@ def ssh_keys_action(module, state):
 
     if state == 'present':
         if target_key == 'absent':
+            changed = True
             data = json.dumps({
                 'name': new_key_name,
                 'default': module.params['default'],
                 'key': module.params['ssh_key']
             })
-            target_key = requests_wrapper(SSH_API, method='POST', data=data).json()
-            changed = True
+            if not module.check_mode:
+                target_key = requests_wrapper(SSH_API, method='POST', data=data).json()
         else:
+            check_immutable_arguments(IMMUTABLE_ARGUMENTS, target_key, module)
             if target_key['default'] != module.params['default']:
+                changed = True
                 data = json.dumps({
                     'name': target_key['name'],
                     'default': module.params['default']
                 })
-                target_key = requests_wrapper(SSH_API + target_key['id'], method='PUT', data=data).json()
-                changed = True
+                if not module.check_mode:
+                    target_key = requests_wrapper(SSH_API + target_key['id'], method='PUT', data=data).json()
 
     if state == 'absent' and target_key != 'absent':
+        changed = True
         data = json.dumps({
             'ssh_key_id': target_key['id']
         })
-        target_key = requests_wrapper(SSH_API + target_key['id'], method='DELETE', data=data).json()
-        changed = True
+        if not module.check_mode:
+            target_key = requests_wrapper(SSH_API + target_key['id'], method='DELETE', data=data).json()
+
+    if target_key == 'absent':
+        target_key = 'The SSH Key [%s]' % new_key_name + ' is absent'
 
     return{
         'changed': changed,
@@ -174,7 +184,8 @@ def main():
             ssh_key=dict(no_log=True),
             state=dict(choices=ALLOWED_STATES, default='present')
         ),
-        required_if=[["state", "absent", ["name"]]]
+        required_if=[["state", "absent", ["name"]]],
+        supports_check_mode=True
     )
 
     if not HAS_REQUESTS:
